+++ llm/providers/openai/mcp_test.go
@@ -1,4 +1,4 @@
-package openairesponses
+package openai
 
 import (
 	"testing"
