+++ llm/providers/openai/tool_image_generation_test.go
@@ -1,4 +1,4 @@
-package openairesponses
+package openai
 
 import (
 	"testing"
